generalization for troop skill perks
added a generalized form for all current career using a perk increasing a units skill. Also added this new form to the ironbreaker

fixed a typo in the choice groups of the knight of the old world.

fixed some (BUT Not all!) internationalizations, will be finished by the end of the pr.